move reaction handler to review cog
